fix(efi): the GOP framebuffer address stops being truncated - T-11
Item 8. buildefi.sh targets x86_64-unknown-windows, which is LLP64:
`unsigned long` is FOUR bytes there while a pointer is eight. The GOP
base is a UINT64 and every hop of the chain was `unsigned long`, so the
top half was thrown away by a cast the programmer wrote on purpose -
which is why the four -Werror flags buildefi.sh carries for this exact
bug class are silent on it. They catch pointer<->int.

IT IS SIX DECLARATIONS, NOT THREE. The task list names efi.c's fb_addr,
console_init_efi and fb_setup's parameter. Between the last two sits
console_init_fb, declared in three places (efi.c, console.c prototype,
console.c definition). Changing a subset is worse than changing none:
the declaration and the definition would then disagree about the size
of a register argument, which is undefined behaviour rather than a
diagnostic.

AND ONE OF THEM IS ASSEMBLY, which nothing in C would have caught.
raw_entry.S calls console_init_fb by hand for the raw-bootloader path -
five dwords pushed, `add $20, %esp`. Widening the first argument to
eight bytes silently shifts every argument after it one slot. It pushes
a zero high dword and adds 24 now, and verify-raw.sh proves it on both
legs.

(The recon claimed intel.c declares console_init_fb with `uptr`, giving
a live ABI mismatch in the EFI build today. It does not - intel.c
mentions it in a comment only. Checked rather than taken on trust.)

THE REGRESSION GUARD IS PROVEN, NOT ASSERTED. _Static_assert(sizeof(addr)
>= sizeof(void *)) at all three hops. Narrowing fb_setup back to
`unsigned long` and running buildefi.sh:

  fb.c:637:20: error: static assertion failed due to requirement
  'sizeof (addr) >= sizeof(void *)': fb_setup's address must be able to
  hold a pointer

The narrowing is spelled out once, in fb_uptr, at the single place an
address becomes a pointer - casting a 64-bit integer straight to a
32-bit pointer is a warning on the 32-bit builds and an error under
-Werror=int-to-pointer-cast.

AND verify-efi.sh NOW EXISTS, which is P-1. Three documents told you to
run it - PLATFORM-PROMPT §1.4, item 8's own gate, desktop-wiring.md -
and it was not in the repo. What was there: verify.sh (32-bit BIOS),
verify-raw.sh (our bootloader), verify-iso.sh (BIOS and UEFI, but both
through GRUB), and buildefi.sh, which builds BOOTX64.EFI and which
NOTHING BOOTED.

That gap is why this bug survived. verify-iso.sh's UEFI leg boots the
MULTIBOOT kernel under OVMF; efi.c is not in that binary. The file
containing the truncation - and the whole firmware-handoff path the
ThinkPad actually takes, with no bootloader in between - was compiled by
the build system and executed by nothing. It is also the only gate where
fb.c's SIMD is live on the path that ships.

It builds a FAT ESP with mtools (no root, no loop devices, the way
mkusb.sh does), boots it under OVMF, and asserts the kernel started AS A
UEFI APPLICATION rather than via multiboot, got a framebuffer, reached
the prompt, opened compositor windows, and that fb_setup did not refuse
the mode. Polls for its marker; no fixed wall-clock wait.

  ok    framebuffer console, 120x37, compositor: 3 windows
  EFI application gate green

Latent rather than active, and that is the point: a GOP base is normally
a PCI BAR in the 32-bit MMIO window, which is why QEMU and OVMF never
showed it. Firmware placing it above 4 GiB gives a black screen or
writes into whatever lives at the truncated address.

Gates, all green:
  verify-efi.sh  NEW - native UEFI application, reaches the compositor
  verify.sh      transcript matches golden.txt, byte-identical
  verify-raw.sh  both legs - the assembly stack change
  verify-iso.sh / verify-sources.sh
  wmtest/inputtest/tritest 0 failures; fbbench hashes unchanged
  probe-smp.py   4 cores, 0 of 961608 pixels differ

+++ kernel/efi.c
@@ -198,9 +198,9 @@ static void capture_acpi(efi_system_table *st)
 
 /* ---- what the kernel proper needs --------------------------------------- */
 void serial_init(void);
-void console_init_fb(unsigned long addr, unsigned int pitch, unsigned int width,
+void console_init_fb(unsigned long long addr, unsigned int pitch, unsigned int width,
                      unsigned int height, unsigned int bpp);
-void console_init_efi(unsigned long addr, unsigned int pitch, unsigned int width,
+void console_init_efi(unsigned long long addr, unsigned int pitch, unsigned int width,
                       unsigned int height, unsigned int bpp);
 void console_init(unsigned long mb_addr);
 int  main(void);
@@ -209,7 +209,13 @@ void kernel_done(void);
 /* Remember the mode across ExitBootServices. Everything the firmware owns
  * becomes invalid the moment that call returns, so anything we still need has
  * to be plain numbers in our own memory before then. */
-static unsigned long fb_addr;
+/* THE GOP BASE IS 64 BITS AND SO IS THIS. It used to be `unsigned long`, and
+ * buildefi.sh targets x86_64-unknown-windows, which is LLP64: `unsigned long`
+ * is FOUR bytes there while a pointer is eight. The explicit cast below then
+ * threw away the top half of a UINT64 silently, and the four -Werror flags in
+ * buildefi.sh do not catch it - they catch pointer<->int, and this is a UINT64
+ * narrowed by a cast the programmer wrote on purpose. T-11. */
+static unsigned long long fb_addr;
 static unsigned int  fb_w, fb_h, fb_pitch_bytes;
 
 MS efi_status efi_main(efi_handle image, efi_system_table *st)
@@ -230,7 +236,7 @@ MS efi_status efi_main(efi_handle image, efi_system_table *st)
     efi_gop *gop = 0;
     if (locate_protocol(&GOP_GUID, 0, (void **)&gop) == EFI_SUCCESS && gop && gop->mode) {
         efi_gop_mode_info *mi = gop->mode->info;
-        fb_addr        = (unsigned long)gop->mode->framebuffer_base;
+        fb_addr        = gop->mode->framebuffer_base;
         fb_w           = mi->horizontal_resolution;
         fb_h           = mi->vertical_resolution;
         /* the stride is given in pixels; our renderer wants bytes */
